+++ fedsim/__init__.py
@@ -1,7 +1,7 @@
 """ a simple Federated Learning simulator!
 """
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 
 import logging
 import os