+++ fedsim/distributed/decentralized/training/__init__.py
@@ -1,12 +1,7 @@
 import os
 
-from .basic_data_manager import BasicDataManager
-from .data_manager import DataManager
-
 __all__ = [
     module[:-3]
     for module in os.listdir(os.path.dirname(__file__))
     if module != "__init__.py" and module[-3:] == ".py"
 ]
-
-__all__ += ["BasicDataManager", "DataManager"]