+++ scripts/__init__.py
@@ -2,7 +2,7 @@
 
 """
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 
 import logging
 import os